+++ R/model.R
@@ -222,6 +222,52 @@ tabnet_config <- function(batch_size = 1024^2,
     skip_importance = skip_importance
   )
 }
+# thanks to https://github.com/tdhock/2023-res-baz-az/blob/main/HOCKING-slides-TRUG.R
+nn_aum_loss <- torch::nn_module(
+  "nn_aum_loss",
+  inherit = torch::nn_mse_loss,
+  initialize = function(){
+    super$initialize()
+  },
+  forward = function(pred_tensor, label_tensor){
+    is_positive = label_tensor == 2L
+    is_negative = is_positive$bitwise_not()
+    # manage case when prediction error is null
+    if(as.logical(torch::torch_sum(is_positive) == 0) || as.logical(torch::torch_sum(is_negative) == 0)){
+      return(torch::torch_sum(pred_tensor*0))
+    }
+    # nominal case
+    fn_diff = torch::torch_where(is_positive, -1, 0)
+    fp_diff = torch::torch_where(is_positive, 0, 1)
+    thresh_tensor = -pred_tensor$flatten()
+    sorted_indices = torch::torch_argsort(thresh_tensor)
+    fp_denom = torch::torch_sum(is_negative) #or 1 for AUM based on count instead of rate
+    fn_denom = torch::torch_sum(is_positive) #or 1 for AUM based on count instead of rate
+    sorted_fp_cum = fp_diff[sorted_indices]$cumsum(dim=1)/fp_denom
+    sorted_fn_cum = -fn_diff[sorted_indices]$flip(1)$cumsum(dim=1)$flip(1)/fn_denom
+    sorted_thresh = thresh_tensor[sorted_indices]
+    sorted_is_diff = sorted_thresh$diff() != 0
+    sorted_fp_end = torch::torch_cat(c(sorted_is_diff, torch::torch_tensor(TRUE)))
+    sorted_fn_end = torch::torch_cat(c(torch::torch_tensor(TRUE), sorted_is_diff))
+    uniq_thresh = sorted_thresh[sorted_fp_end]
+    uniq_fp_after = sorted_fp_cum[sorted_fp_end]
+    uniq_fn_before = sorted_fn_cum[sorted_fn_end]
+    FPR = torch::torch_cat(c(torch::torch_tensor(0.0), uniq_fp_after))
+    FNR = torch::torch_cat(c(uniq_fn_before, torch::torch_tensor(0.0)))
+    roc = list(
+      FPR=FPR,
+      FNR=FNR,
+      TPR=1 - FNR,
+      "min(FPR,FNR)"=torch::torch_minimum(FPR, FNR),
+      min_constant=torch::torch_cat(c(torch::torch_tensor(-Inf), uniq_thresh)),
+      max_constant=torch::torch_cat(c(uniq_thresh, torch::torch_tensor(Inf))))
+    min_FPR_FNR = roc[["min(FPR,FNR)"]][2:-2]
+    constant_diff = roc$min_constant[2:N]$diff()
+    torch::torch_sum(min_FPR_FNR * constant_diff)
+    
+    
+  }
+)
 
 get_constr_output <- function(x, R) {
     # MCM of the prediction given the hierarchy constraint expressed in the matrix R """