+++ R/pretraining.R
@@ -43,20 +43,31 @@ transpose_metrics <- function(metrics) {
   out[-1]
 }
 
-unsupervised_loss <- function(y_pred, embedded_x, obfuscation_mask, eps = 1e-9) {
 
-  errors <- y_pred - embedded_x
-  reconstruction_errors <- torch::torch_mul(errors, obfuscation_mask)^2
-  batch_stds <- torch::torch_std(embedded_x, dim=1)^2 + eps
+nn_unsupervised_loss <- torch::nn_module(
+  "nn_unsupervised_loss",
+  inherit = torch::nn_cross_entropy_loss,
+  
+  initialize = function(eps = 1e-9){
+    super$initialize()
+    self$eps = eps
+  },
+  
+  forward = function(y_pred, embedded_x, obfuscation_mask){
+    errors <- y_pred - embedded_x
+    reconstruction_errors <- torch::torch_mul(errors, obfuscation_mask) ^ 2
+    batch_stds <- torch::torch_std(embedded_x, dim = 1) ^ 2 + self$eps
+    
+    # compute the number of obfuscated variables to reconstruct
+    nb_reconstructed_variables <- torch::torch_sum(obfuscation_mask, dim = 2)
+    
+    # take the mean of the reconstructed variable errors
+    features_loss <- torch::torch_matmul(reconstruction_errors, 1 / batch_stds) / (nb_reconstructed_variables +  self$eps)
+    loss <- torch::torch_mean(features_loss, dim = 1)
+    loss
+  }
+)
 
-  # compute the number of obfuscated variables to reconstruct
-  nb_reconstructed_variables <- torch::torch_sum(obfuscation_mask, dim=2)
-
-  # take the mean of the reconstructed variable errors
-  features_loss <- torch::torch_matmul(reconstruction_errors, 1/batch_stds) / (nb_reconstructed_variables + eps)
-  loss <- torch::torch_mean(features_loss)
-  loss
-}
 
 tabnet_train_unsupervised <- function(x, config = tabnet_config(), epoch_shift = 0L) {
   torch::torch_manual_seed(sample.int(1e6, 1))
@@ -103,7 +114,7 @@ tabnet_train_unsupervised <- function(x, config = tabnet_config(), epoch_shift =
   )
 
   # resolve loss (shortcutted from config)
-  config$loss_fn <- unsupervised_loss
+  config$loss_fn <- nn_unsupervised_loss()
 
   # create network
   network <- tabnet_pretrainer(