+++ NEWS.md
@@ -1,3 +1,5 @@
+# tabnet (development version)
+
 # tabnet 0.7.0
 
 ## Bugfixes